+++ database/seeders/DatabaseSeeder.php
@@ -2,16 +2,17 @@
 
 declare(strict_types=1);
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
-class UsersSeeder extends Seeder
+class DatabaseSeeder extends Seeder
 {
     /**
      * Run the database seeds.
-     * Run "php artisan db:seed --class=UsersSeeder".
      */
     public function run(): void
     {
-        \App\User::factory()->count(10)->create();
+        $this->call(TeamsSeeder::class);
     }
 }
