+++ database/seeders/DuesTransactionsSeeder.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class DuesTransactionsSeeder extends Seeder
