+++ database/seeders/FasetVisitsSeeder.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class FasetVisitsSeeder extends Seeder
