+++ database/seeders/UsersSeeder.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Database\Seeders;
+
+use Illuminate\Database\Seeder;
+
+class UsersSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     * Run "php artisan db:seed --class=UsersSeeder".
+     */
+    public function run(): void
+    {
+        \App\User::factory()->count(10)->create();
+    }
+}
