--- database/seeds/DatabaseSeeder.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Illuminate\Database\Seeder;
-
-class DatabaseSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     */
-    public function run(): void
-    {
-        $this->call(TeamsSeeder::class);
-    }
-}
